Add other_occupation to user schema and move picker items of the profile screen to separate component
